second

+++ app/Http/Controllers/Auth/Otp/OtpValidationController.php
@@ -8,19 +8,19 @@ use GuzzleHttp\Client;
 use App\Models\Setting;
 use App\Models\UserOtp;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Http;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Validation\ValidationException;
 use App\Http\Requests\Auth\Otp\OtpValidationRequest;
+use App\Http\Requests\Auth\Register\ProfileUpdateRequest;
 
 class OtpValidationController extends Controller
 {
     public function index($userId)
     {
 
-        $userId = decrypt($userId);
         $now = now();
         $dataUserOtp = UserOtp::query()->where('user_id', $userId)->latest()->first();
-        // dd($dataUserOtp);
 
         return view('auth.register.otp', compact('userId', 'now', 'dataUserOtp'));
     }
@@ -58,28 +58,18 @@ class OtpValidationController extends Controller
             'message' => "Halo $userName\nKami dari $communityName 🚖\n\nMasukkan Kode OTP berikut untuk verifikasi:\n\nKode OTP : $dataUserOtp->otp\n*Expired dalam waktu 5 menit, Jam $expiredHour:$expiredMinute:$expiredSecond\n\nCatatan* Jika Akang/Eteh merasa tidak melakukan pendaftaran apapun, silahkan abaikan pesan ini. Terima kasih🙏"
         ];
 
-        $client = new Client();
-        $url = $endPoint;
 
-        try {
-            $response = $client->post($url, [
-                'json' => $parameter
-            ]);
+        $http = Http::post($endPoint, $parameter);
 
+        if ($http->ok()) {
             if ($updateUserOtp) {
                 // Use session message for success
                 return back()->with(['success' => "Berhasil mengirim ulang OTP"]);
             } else {
-                // Use session message for error
-                return back()->with(['error' => "Silahkan coba lagi setelah beberapa saat"]);
+                return back()->with(['error' => "Gagal mengirim ulang OTP, periksa koneksi internet anda dan coba lagi."]);
             }
-        } catch (\Exception $e) {
-            $errorMessage = $e->getMessage();
-            // Use session message for exception handling
-            return back()->with(['error' => "Gagal mengirim ulang OTP, periksa koneksi internet anda dan coba lagi."]);
         }
     }
-
     public function verification(OtpValidationRequest $request, $userId)
     {
         $now = now();
@@ -95,7 +85,7 @@ class OtpValidationController extends Controller
             $setting = Setting::firstOrFail();
             $communityName = $setting->community_name;
 
-            // Send message to whatsapp
+            // Send message to WhatsApp
             $endPoint = $setting->endpoint;
             $apiKey = $setting->api_key;
             $sender = $setting->sender;
@@ -110,25 +100,57 @@ class OtpValidationController extends Controller
                 'message' => "Halo, $userName\nKami dari $communityName 🚖\n\nSelamat akun anda $userName berhasil dibuat.\nUntuk informasi lebih lanjut silahkan kunjungi $appName"
             ];
 
-            $client = new Client();
-            $url = $endPoint;
+            $http = Http::post($endPoint, $parameter);
 
-            try {
-                $response = $client->post($url, [
-                    'json' => $parameter
-                ]);
-
+            if ($http->ok()) {
                 $checkOtp->update([
-                    'isVerified' => '1',
+                    'isVerified' => 1,
                 ]);
+
                 $user = User::findOrFail($userId);
+                $user->update([
+                    'phone_verified_at' => now(),
+                ]);
+
                 $checkOtp->delete();
-                Auth::login($user);
-            } catch (\Exception $e) {
-                $errorMessage = $e->getMessage();
+
+                // Redirect to the update profile page
+                return redirect()->route('register.verificationOtp.update', ['userId' => $userId])->with('success', 'Akun anda berhasil dibuat, silahkan lengkapi data diri anda. Terima kasih.');
+            } else {
                 return back()->with(['failed' => "Periksa koneksi internet anda dan coba beberapa saat lagi."]);
             }
-            return redirect()->route('auth.login')->with(['success' => $checkOtp->user->name . "Berhasil Login"]);
         }
+    }
+
+    public function update($userId)
+    {
+        $user = User::findOrFail($userId);
+
+        return view('auth.register.profile', compact('user'));
+    }
+
+    public function verified(ProfileUpdateRequest $request, $userId)
+    {
+
+        $user = User::findOrFail($userId);
+        $user->update([
+            'fullname' => $request->fullname,
+            'email' => $request->email,
+            'gender' => $request->gender,
+            'date_birth' => $request->date_birth,
+            'address' => $request->address,
+            'province' => $request->province,
+            'city' => $request->city,
+            'district' => $request->district,
+            'postal_code' => $request->postal_code,
+            'status' =>  'waiting',
+        ]);
+
+        if (!$user->hasVerifiedEmail()) {
+            $user->sendEmailVerificationNotification();
+            return redirect()->route('auth.login')->with('success', 'Profil Anda berhasil diperbarui. Silahkan verifikasi alamat email Anda.');
+        }
+
+        return redirect()->route('auth.login')->with('success', 'Profil Anda berhasil diperbarui. Silahkan verifikasi alamat email Anda.');
     }
 }
